// src/dto/user.dto.ts
export class UserDTO {
    id?: number;
    name?: string;
    steamid64?: string;
    sid?: string;
    avatar?: string;
    profile_url?: string;
    real_name?: string;
    date_of_birth?: number;
    last_online?: number;
    balance?: number;
    virtual_balance?: number;
    discord?: string;
    friendTrades?: boolean;
    // group_id?: Privileges;
    vkURL?: string;
}